add proper error management
